+++ HISTORY.md
@@ -1,6 +1,7 @@
 unreleased
 ==========
 
+  * remove support for non-`res` argument
   * deps: ee-first@1.0.5
 
 1.2.2 / 2014-06-10