+++ HISTORY.md
@@ -1,6 +1,7 @@
 unreleased
 ==========
 
+  * add `isFinished` function
   * remove support for non-`res` argument
   * support both `req` and `res` as arguments
   * deps: ee-first@1.0.5