+++ HISTORY.md
@@ -1,5 +1,5 @@
-unreleased
-==========
+2.0.0 / 2014-08-16
+==================
 
   * Add `isFinished` function
   * Move to `jshttp` organization